backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
pypdf
python-docx
python-multipart
openai>=1.0.0
spacy
transformers
torch
accelerate
